+++ 1.20.4/Forge/src/main/java/fuzs/puzzleslib/forge/impl/capability/data/ForgeBlockEntityCapabilityKey.java
@@ -4,14 +4,15 @@ import fuzs.puzzleslib.api.capability.v3.data.BlockEntityCapabilityKey;
 import fuzs.puzzleslib.api.capability.v3.data.CapabilityComponent;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.level.block.entity.BlockEntity;
+import net.minecraftforge.common.capabilities.Capability;
 
 import java.util.function.Predicate;
 import java.util.function.Supplier;
 
 public class ForgeBlockEntityCapabilityKey<T extends BlockEntity, C extends CapabilityComponent<T>> extends ForgeCapabilityKey<T, C> implements BlockEntityCapabilityKey<T, C> {
 
-    public ForgeBlockEntityCapabilityKey(ResourceLocation identifier, CapabilityTokenFactory<T, C> tokenFactory, Predicate<Object> filter, Supplier<C> capabilityFactory) {
-        super(identifier, tokenFactory, filter, capabilityFactory);
+    public ForgeBlockEntityCapabilityKey(ResourceLocation identifier, Capability<C> capability, Predicate<Object> filter, Supplier<C> capabilityFactory) {
+        super(identifier, capability, filter, capabilityFactory);
     }
 
     @Override